pyproject.toml
[project]
name = "medication-assistant"
version = "0.1.0"
description = "智能用药安全助手"
requires-python = ">=3.11, <3.13"
dependencies = [
    # LangChain (既然已经是 1.2.x 了，就先这样用着)
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=1.0.0",
    # 向量数据库
    "langchain-chroma>=0.1.4",
    "chromadb>=0.5.20",
    # OpenAI
    "openai>=1.54.4",
    # Web框架
    "streamlit>=1.40.1",
    "fastapi>=0.115.5",
    "uvicorn>=0.32.1",
    # 数据处理
    "pandas>=2.2.3",
    "numpy<2.0.0",
    # 工具
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.6.1",
    # 日志和监控
    "loguru>=0.7.2",
    "dashscope>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "black>=24.10.0",
    "ruff>=0.8.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
sources = ["src"]